Trees now read data
This is the first stab at getting trees to read data. Right now setpoints are hardcded and the data still only works with floats. the next step is to integrate the tree creation with the game by simply increasing the score everytime the user adds a node to the tree. I'm saving this commit just in case the next implementation goes haywire.